Adding unit tests for Context Value
