+++ Eklee.Azure.Functions.GraphQl/AssertAction.cs
@@ -1,0 +1,11 @@
+﻿namespace Eklee.Azure.Functions.GraphQl
+{
+	public enum AssertAction
+	{
+		BatchCreate,
+		Create,
+		Update,
+		Delete,
+		DeleteAll
+	}
+}
