+++ Eklee.Azure.Functions.GraphQl/Comparisons.cs
@@ -1,0 +1,15 @@
+﻿namespace Eklee.Azure.Functions.GraphQl
+{
+	public enum Comparisons
+	{
+		Equal,
+		StringStartsWith,
+		StringEndsWith,
+		StringContains,
+		NotEqual,
+		GreaterThan,
+		GreaterEqualThan,
+		LessThan,
+		LessEqualThan
+	}
+}
